슈팅게임 배경, 피격 완성

+++ GameFramework/Sample.h
@@ -18,7 +18,7 @@ private:
 	InputState inputState;
 
 private:
-	int posX;
-	int posY;
-	int moveSpeed;
+	float posX;
+	float posY;
+	float moveSpeed;
 };
